+++ src/utils/paths/utils.go
@@ -7,11 +7,15 @@ import (
 
 // IsAncesterOf returns true if ancestor is an ancestor of path.
 func IsAncesterOf(ancestor string, path string) bool {
-	ancestor, err := filepath.Abs(ancestor)
+	var err error
+
+	ancestor = filepath.FromSlash(ancestor)
+	ancestor, err = filepath.Abs(ancestor)
 	if err != nil {
 		return false
 	}
 
+	path = filepath.FromSlash(path)
 	path, err = filepath.Abs(path)
 	if err != nil {
 		return false
@@ -25,5 +29,24 @@ func IsAncesterOf(ancestor string, path string) bool {
 	// Convert to slash for Windows.
 	relativePath = filepath.ToSlash(relativePath)
 
-	return !strings.HasPrefix(relativePath, "../") && relativePath != ".."
+	return !strings.HasPrefix(relativePath, "../") && relativePath != ".." && !IsIdentical(ancestor, path)
+}
+
+// IsIdentical returns true if path1 and path2 are identical.
+func IsIdentical(path1 string, path2 string) bool {
+	var err error
+
+	path1 = filepath.FromSlash(path1)
+	path1, err = filepath.Abs(path1)
+	if err != nil {
+		return false
+	}
+
+	path2 = filepath.FromSlash(path2)
+	path2, err = filepath.Abs(path2)
+	if err != nil {
+		return false
+	}
+
+	return path1 == path2
 }